Add "barebones" InteractionContext, useful if you are storing the interaction token in a database but you want to reply to it
